[pkg] Upgrade TypeScript to 3.7
