+++ package.json
@@ -89,13 +89,13 @@
     "@anansi/eslint-config": "^0.6.6",
     "@babel/cli": "^7.7.0",
     "@babel/core": "^7.7.0",
-    "@testing-library/react": "^9.3.0",
-    "@testing-library/react-hooks": "^3.1.1",
-    "@types/jest": "^24.0.20",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/react-hooks": "^3.2.1",
+    "@types/jest": "^24.0.22",
     "@types/lodash": "^4.14.144",
     "@types/react": "^16.9.11",
-    "@typescript-eslint/eslint-plugin": "^2.6.0",
-    "@typescript-eslint/parser": "^2.6.0",
+    "@typescript-eslint/eslint-plugin": "^2.6.1",
+    "@typescript-eslint/parser": "^2.6.1",
     "@zerollup/ts-transform-paths": "^1.7.3",
     "auto-changelog": "^1.16.1",
     "babel-core": "^7.0.0-bridge.0",
@@ -126,7 +126,7 @@
     "scheduler": "^0.17.0",
     "ts-jest": "^24.1.0",
     "ttypescript": "^1.5.7",
-    "typescript": "^3.6.4"
+    "typescript": "^3.7.2"
   },
   "dependencies": {
     "@babel/runtime": "^7.7.1",
