+++ src/resource/index.ts
@@ -1,5 +1,6 @@
 import Resource from './Resource';
 import SimpleResource from './SimpleResource';
+export * from './shapes';
 export * from './types';
 export * from './normal';
